refactor: store에 전역상태 모으기 및 적용된 경로 수정
